Counter is working better.
Still need some improvements...
