+++ gameproject/src/main/java/com/qatrain/janushgame/MainApp.java
@@ -1,6 +1,6 @@
 package com.qatrain.janushgame;
 
-import com.qatrain.janushgame.model.Counter;
+import com.qatrain.janushgame.model.Game;
 
 /**
  * This class starts the whole game.
@@ -9,7 +9,8 @@ public class MainApp {
 
     public static void main(String... args) {
 
-        new Counter();
+        Game game = new Game();
+        game.play();
 
     }
 }
